+++ src/triggers/sequence.ts
@@ -1,5 +1,8 @@
 import { IPlayer, ITrigger } from '../interfaces.js';
 
+const NUMBER_REGEX = /^\d*(\.\d+)?$/
+const MUTATION_OBSERVER_CONFIG = { attributes: true, childList: false, subtree: false };
+
 /**
  * Sequence player.
  * 
@@ -14,12 +17,21 @@ export class Sequence implements ITrigger {
     protected frameDelayFirst: number | null = null;
     protected frameDelayLast: number | null = null;
     protected timer: any;
+    protected observer: MutationObserver;
 
     constructor(
         protected element: HTMLElement,
         protected targetElement: HTMLElement,
         protected player: IPlayer,
     ) {
+        this.observer = new MutationObserver((mutationList, observer) => {
+            for (const mutation of mutationList) {
+                if (mutation.type === "attributes" && mutation.attributeName === 'sequence') {
+                    this.reset();
+                    this.step();
+                }
+            }
+        });
     }
 
     onReady() {
@@ -36,16 +48,30 @@ export class Sequence implements ITrigger {
     }
 
     onConnected() {
+        this.observer.observe(this.element, MUTATION_OBSERVER_CONFIG);
     }
 
     onDisconnected() {
+        this.observer.disconnect();
+
         if (this.timer) {
             clearTimeout(this.timer);
             this.timer = null;
         }
     }
 
-    takeNextStep() {
+    reset() {
+        this.player.pause();
+        this.sequenceIndex = 0;
+        this.frameState = this.frameDelayFirst = this.frameDelayLast = null;
+
+        if (this.timer) {
+            clearTimeout(this.timer);
+            this.timer = null;
+        }
+    }
+
+    takeStep() {
         const steps = this.sequence.split(',');
 
         const step = steps[this.sequenceIndex];
@@ -55,16 +81,12 @@ export class Sequence implements ITrigger {
             this.sequenceIndex = 0;
         }
 
-        return step;
-    }
-
-    step() {
-        const numberRegex = /^\d*(\.\d+)?$/
-
-        const step = this.takeNextStep();
-
         const [action, ...params] = step.split(':');
 
+        return { action, params };
+    }
+
+    handleStep(action: string, params: string[]) {
         if (action === 'play') {
             if (this.frameState) {
                 this.player.state = this.frameState;
@@ -88,7 +110,7 @@ export class Sequence implements ITrigger {
             }, this.frameDelayFirst || 0);
         } else if (action === 'frame') {
             let frameIndex = 0;
-            if (params.length && params[0].match(numberRegex)) {
+            if (params.length && params[0].match(NUMBER_REGEX)) {
                 frameIndex = Math.max(0, Math.min(this.player.frames, +params[0]));
             }
 
@@ -108,7 +130,7 @@ export class Sequence implements ITrigger {
             let value: number | null = null;
 
             for (const param of params) {
-                if (param && param.match(numberRegex)) {
+                if (param && param.match(NUMBER_REGEX)) {
                     value = +param;
                 }
             }
@@ -127,12 +149,24 @@ export class Sequence implements ITrigger {
             }
 
             this.step();
+        } else if (action === 'idle') {
+            // do nothing
         } else {
             throw new Error(`Invalid sequence action: ${action}`);
         }
     }
 
+    step() {
+        const { action, params } = this.takeStep();
+
+        if (!action) {
+            return;
+        }
+
+        this.handleStep(action, params);
+    }
+
     get sequence() {
-        return this.element.getAttribute('sequence') || 'play';
+        return this.element.getAttribute('sequence') || '';
     }
 }